Use SS:GraphBLAS header 3.3.3.  Oops.
Also, add a regression test that makes sures the header file versions match.
