+++ grblas/backends/suitesparse/suitesparse_graphblas_3.3.3.h
@@ -112,10 +112,10 @@
 
 // The version of this implementation, and the GraphBLAS API version:
 /* #define GxB_IMPLEMENTATION_NAME "SuiteSparse:GraphBLAS" */
-/* #define GxB_IMPLEMENTATION_DATE "June 30, 2020" */
+/* #define GxB_IMPLEMENTATION_DATE "JULY 14, 2020" */
 #define GxB_IMPLEMENTATION_MAJOR 3
 #define GxB_IMPLEMENTATION_MINOR 3
-#define GxB_IMPLEMENTATION_SUB   1
+#define GxB_IMPLEMENTATION_SUB   3
 /* #define GxB_SPEC_DATE "Sept 25, 2019" */
 #define GxB_SPEC_MAJOR 1
 #define GxB_SPEC_MINOR 3
